# components/api-server/pkg/api/openapi/api/openapi.yaml
openapi: 3.0.3
info:
  description: HyperShell fleet management API
  license:
    name: Apache-2.0
    url: https://www.apache.org/licenses/LICENSE-2.0.html
  title: HyperShell API
  version: 1.0.0
servers:
- description: localhost
  url: http://localhost:8000
- description: current domain
  url: /
security:
- Bearer: []
...
paths:
...
  /api/hypershell/v1/gateways/{id}:
    delete:
      operationId: deleteGateway
      parameters:
      - description: The id of record
        explode: false
        in: path
        name: id
        required: true
        schema:
          type: string
        style: simple
      responses:
        "204":
          description: Gateway deleted successfully
        "401":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
          description: Auth token is invalid
        "403":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
          description: Unauthorized to perform operation
        "404":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
          description: No gateway with specified id exists
        "500":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
          description: Unexpected error deleting gateway
      security:
      - Bearer: []
      summary: Delete a gateway
    get:
      operationId: getGateway
      parameters:
      - description: The id of record
        explode: false
        in: path
        name: id
        required: true
        schema:
          type: string
        style: simple
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Gateway"
          description: Gateway found by id
        "401":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
          description: Auth token is invalid
        "403":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
          description: Unauthorized to perform operation
        "404":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
          description: No gateway with specified id exists
        "500":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
          description: Unexpected error occurred
      security:
      - Bearer: []
      summary: Get an gateway by id
    patch:
      operationId: updateGateway
      parameters:
      - description: The id of record
        explode: false
        in: path
        name: id
        required: true
        schema:
          type: string
        style: simple
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/GatewayPatchRequest"
        description: Updated gateway data
        required: true
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Gateway"
          description: Gateway updated successfully
        "400":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
          description: Validation errors occurred
        "401":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
          description: Auth token is invalid
        "403":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
          description: Unauthorized to perform operation
        "404":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
          description: No gateway with specified id exists
        "409":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
          description: Gateway already exists
        "500":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
          description: Unexpected error updating gateway
      security:
      - Bearer: []
      summary: Update an gateway
...
components:
...
  schemas:
...
    ObjectReference:
      example:
        updated_at: 2000-01-23T04:56:07.000+00:00
        kind: kind
        created_at: 2000-01-23T04:56:07.000+00:00
        id: id
        href: href
      properties:
        id:
          type: string
        kind:
          type: string
        href:
          type: string
        created_at:
          format: date-time
          type: string
        updated_at:
          format: date-time
          type: string
      type: object
    Error:
      allOf:
      - $ref: "#/components/schemas/ObjectReference"
      - properties:
          code:
            type: string
          reason:
            type: string
          operation_id:
            type: string
        type: object
      example:
        reason: reason
        code: code
        updated_at: 2000-01-23T04:56:07.000+00:00
        kind: kind
        created_at: 2000-01-23T04:56:07.000+00:00
        operation_id: operation_id
        id: id
        href: href
...
    Gateway:
      allOf:
      - $ref: "#/components/schemas/ObjectReference"
      - properties:
          name:
            type: string
          fleet_id:
            type: string
          cluster_id:
            type: string
          release_id:
            type: string
          database_id:
            type: string
          namespace:
            description: API-assigned Kubernetes namespace derived from the Gateway
              identifier
            readOnly: true
            type: string
          external_dns:
            type: string
          tls_mode:
            type: string
          service_type:
            type: string
          status:
            type: string
          phase:
            type: string
          image:
            description: Container image for the gateway deployment
            type: string
          supervisor_image:
            description: Container image for the supervisor sidecar
            type: string
          server_dns_names:
            description: DNS names for TLS certificate SANs
            items:
              type: string
            type: array
          route_address:
            description: External route address populated by the control plane
            readOnly: true
            type: string
          console_address:
            description: Web console address populated by the control plane
            readOnly: true
            type: string
          oidc:
            description: JSON-encoded OIDC authentication configuration (auto-populated
              by Keycloak provisioning)
            readOnly: true
            type: string
          route:
            description: JSON-encoded route configuration
            type: string
          database_config:
            description: JSON-encoded database provisioning configuration
            type: string
          credential_driver:
            description: JSON-encoded credential storage driver configuration
            type: string
          active_sandbox_count:
            description: Number of active (Running or Pending) agent sandboxes observed
              in the gateway namespace by the control plane
            format: int32
            readOnly: true
            type: integer
        required:
        - cluster_id
        - database_id
        - fleet_id
        - name
        - namespace
        - release_id
        type: object
      example:
        release_id: release_id
        server_dns_names:
        - server_dns_names
        - server_dns_names
        created_at: 2000-01-23T04:56:07.000+00:00
        credential_driver: credential_driver
        cluster_id: cluster_id
        updated_at: 2000-01-23T04:56:07.000+00:00
        database_config: database_config
        id: id
        href: href
        fleet_id: fleet_id
        external_dns: external_dns
        phase: phase
        image: image
        active_sandbox_count: 5
        kind: kind
        console_address: console_address
        route_address: route_address
        oidc: oidc
        supervisor_image: supervisor_image
        service_type: service_type
        tls_mode: tls_mode
        route: route
        database_id: database_id
        name: name
        namespace: namespace
        status: status
...
    GatewayPatchRequest:
      example:
        phase: phase
        release_id: release_id
        image: image
        server_dns_names:
        - server_dns_names
        - server_dns_names
        console_address: console_address
        credential_driver: credential_driver
        route_address: route_address
        oidc: oidc
        supervisor_image: supervisor_image
        cluster_id: cluster_id
        service_type: service_type
        tls_mode: tls_mode
        route: route
        database_id: database_id
        database_config: database_config
        name: name
        fleet_id: fleet_id
        external_dns: external_dns
        status: status
      properties:
        name:
          type: string
        fleet_id:
          type: string
        cluster_id:
          type: string
        release_id:
          type: string
        database_id:
          type: string
        external_dns:
          type: string
        tls_mode:
          type: string
        service_type:
          type: string
        status:
          type: string
        phase:
          type: string
        image:
          type: string
        supervisor_image:
          type: string
        server_dns_names:
          items:
            type: string
          type: array
        route_address:
          type: string
        console_address:
          type: string
        oidc:
          type: string
        route:
          type: string
        database_config:
          type: string
        credential_driver:
          type: string
      type: object
...
  securitySchemes:
    Bearer:
      bearerFormat: JWT
      scheme: bearer
      type: http