# components/sdk-go/types/base.go
// Code generated by trex-sdk-generator from openapi.yaml - DO NOT EDIT.
// Source: ../../components/api-server/openapi/openapi.yaml
// Spec SHA256: 4f16ee432c4370397b65ba2fc0f1a28d5b813c0ae26c388667d5684a8b825020
// Generated: 1970-01-01T00:00:00Z

package types

import (
	"strconv"
	"time"
)

type ObjectReference struct {
	ID        string     `json:"id,omitempty"`
	Kind      string     `json:"kind,omitempty"`
	Href      string     `json:"href,omitempty"`
	CreatedAt *time.Time `json:"created_at,omitempty"`
	UpdatedAt *time.Time `json:"updated_at,omitempty"`
}

type ListMeta struct {
	Kind  string `json:"kind"`
	Page  int    `json:"page"`
	Size  int    `json:"size"`
	Total int    `json:"total"`
}

type APIError struct {
	ID          string `json:"id,omitempty"`
	Kind        string `json:"kind,omitempty"`
	Href        string `json:"href,omitempty"`
	Code        string `json:"code"`
	Reason      string `json:"reason"`
	OperationID string `json:"operation_id,omitempty"`
	StatusCode  int    `json:"-"`
}

func (e *APIError) Error() string {
	if e.StatusCode > 0 {
		return "API error " + strconv.Itoa(e.StatusCode) + ": " + e.Code + " - " + e.Reason
	}
	return "API error: " + e.Code + " - " + e.Reason
}