# components/sdk-go/types/gateway.go
// Code generated by trex-sdk-generator from openapi.yaml - DO NOT EDIT.
// Source: ../../components/api-server/openapi/openapi.yaml
// Spec SHA256: 4f16ee432c4370397b65ba2fc0f1a28d5b813c0ae26c388667d5684a8b825020
// Generated: 1970-01-01T00:00:00Z

package types

import (
	"errors"
	"fmt"
)

type Gateway struct {
	ObjectReference

	ActiveSandboxCount int32 `json:"active_sandbox_count,omitempty"`
	ClusterID string `json:"cluster_id"`
	ConsoleAddress string `json:"console_address,omitempty"`
	CredentialDriver string `json:"credential_driver,omitempty"`
	DatabaseConfig string `json:"database_config,omitempty"`
	DatabaseID string `json:"database_id"`
	ExternalDNS string `json:"external_dns,omitempty"`
	FleetID string `json:"fleet_id"`
	Image string `json:"image,omitempty"`
	Name string `json:"name"`
	Namespace string `json:"namespace"`
	Oidc string `json:"oidc,omitempty"`
	Phase string `json:"phase,omitempty"`
	ReleaseID string `json:"release_id"`
	Route string `json:"route,omitempty"`
	RouteAddress string `json:"route_address,omitempty"`
	ServerDNSNames string `json:"server_dns_names,omitempty"`
	ServiceType string `json:"service_type,omitempty"`
	Status string `json:"status,omitempty"`
	SupervisorImage string `json:"supervisor_image,omitempty"`
	TLSMode string `json:"tls_mode,omitempty"`
}

type GatewayList struct {
	ListMeta
	Items []Gateway `json:"items"`
}

func (l *GatewayList) GetItems() []Gateway { return l.Items }
func (l *GatewayList) GetTotal() int                  { return l.Total }
func (l *GatewayList) GetPage() int                   { return l.Page }
func (l *GatewayList) GetSize() int                   { return l.Size }

type GatewayBuilder struct {
	resource Gateway
	errors   []error
}

func NewGatewayBuilder() *GatewayBuilder {
	return &GatewayBuilder{}
}


func (b *GatewayBuilder) ClusterID(v string) *GatewayBuilder {
	b.resource.ClusterID = v
	return b
}

func (b *GatewayBuilder) CredentialDriver(v string) *GatewayBuilder {
	b.resource.CredentialDriver = v
	return b
}

func (b *GatewayBuilder) DatabaseConfig(v string) *GatewayBuilder {
	b.resource.DatabaseConfig = v
	return b
}

func (b *GatewayBuilder) DatabaseID(v string) *GatewayBuilder {
	b.resource.DatabaseID = v
	return b
}

func (b *GatewayBuilder) ExternalDNS(v string) *GatewayBuilder {
	b.resource.ExternalDNS = v
	return b
}

func (b *GatewayBuilder) FleetID(v string) *GatewayBuilder {
	b.resource.FleetID = v
	return b
}

func (b *GatewayBuilder) Image(v string) *GatewayBuilder {
	b.resource.Image = v
	return b
}

func (b *GatewayBuilder) Name(v string) *GatewayBuilder {
	b.resource.Name = v
	return b
}

func (b *GatewayBuilder) Phase(v string) *GatewayBuilder {
	b.resource.Phase = v
	return b
}

func (b *GatewayBuilder) ReleaseID(v string) *GatewayBuilder {
	b.resource.ReleaseID = v
	return b
}

func (b *GatewayBuilder) Route(v string) *GatewayBuilder {
	b.resource.Route = v
	return b
}

func (b *GatewayBuilder) ServerDNSNames(v string) *GatewayBuilder {
	b.resource.ServerDNSNames = v
	return b
}

func (b *GatewayBuilder) ServiceType(v string) *GatewayBuilder {
	b.resource.ServiceType = v
	return b
}

func (b *GatewayBuilder) Status(v string) *GatewayBuilder {
	b.resource.Status = v
	return b
}

func (b *GatewayBuilder) SupervisorImage(v string) *GatewayBuilder {
	b.resource.SupervisorImage = v
	return b
}

func (b *GatewayBuilder) TLSMode(v string) *GatewayBuilder {
	b.resource.TLSMode = v
	return b
}


func (b *GatewayBuilder) Build() (*Gateway, error) {
	if b.resource.ClusterID == "" {
		b.errors = append(b.errors, fmt.Errorf("cluster_id is required"))
	}
	if b.resource.DatabaseID == "" {
		b.errors = append(b.errors, fmt.Errorf("database_id is required"))
	}
	if b.resource.FleetID == "" {
		b.errors = append(b.errors, fmt.Errorf("fleet_id is required"))
	}
	if b.resource.Name == "" {
		b.errors = append(b.errors, fmt.Errorf("name is required"))
	}
	if b.resource.Namespace == "" {
		b.errors = append(b.errors, fmt.Errorf("namespace is required"))
	}
	if b.resource.ReleaseID == "" {
		b.errors = append(b.errors, fmt.Errorf("release_id is required"))
	}
	if len(b.errors) > 0 {
		return nil, fmt.Errorf("validation failed: %w", errors.Join(b.errors...))
	}
	return &b.resource, nil
}

type GatewayPatchBuilder struct {
	patch map[string]any
}

func NewGatewayPatchBuilder() *GatewayPatchBuilder {
	return &GatewayPatchBuilder{patch: make(map[string]any)}
}


func (b *GatewayPatchBuilder) ClusterID(v string) *GatewayPatchBuilder {
	b.patch["cluster_id"] = v
	return b
}

func (b *GatewayPatchBuilder) ConsoleAddress(v string) *GatewayPatchBuilder {
	b.patch["console_address"] = v
	return b
}

func (b *GatewayPatchBuilder) CredentialDriver(v string) *GatewayPatchBuilder {
	b.patch["credential_driver"] = v
	return b
}

func (b *GatewayPatchBuilder) DatabaseConfig(v string) *GatewayPatchBuilder {
	b.patch["database_config"] = v
	return b
}

func (b *GatewayPatchBuilder) DatabaseID(v string) *GatewayPatchBuilder {
	b.patch["database_id"] = v
	return b
}

func (b *GatewayPatchBuilder) ExternalDNS(v string) *GatewayPatchBuilder {
	b.patch["external_dns"] = v
	return b
}

func (b *GatewayPatchBuilder) FleetID(v string) *GatewayPatchBuilder {
	b.patch["fleet_id"] = v
	return b
}

func (b *GatewayPatchBuilder) Image(v string) *GatewayPatchBuilder {
	b.patch["image"] = v
	return b
}

func (b *GatewayPatchBuilder) Name(v string) *GatewayPatchBuilder {
	b.patch["name"] = v
	return b
}

func (b *GatewayPatchBuilder) Oidc(v string) *GatewayPatchBuilder {
	b.patch["oidc"] = v
	return b
}

func (b *GatewayPatchBuilder) Phase(v string) *GatewayPatchBuilder {
	b.patch["phase"] = v
	return b
}

func (b *GatewayPatchBuilder) ReleaseID(v string) *GatewayPatchBuilder {
	b.patch["release_id"] = v
	return b
}

func (b *GatewayPatchBuilder) Route(v string) *GatewayPatchBuilder {
	b.patch["route"] = v
	return b
}

func (b *GatewayPatchBuilder) RouteAddress(v string) *GatewayPatchBuilder {
	b.patch["route_address"] = v
	return b
}

func (b *GatewayPatchBuilder) ServerDNSNames(v string) *GatewayPatchBuilder {
	b.patch["server_dns_names"] = v
	return b
}

func (b *GatewayPatchBuilder) ServiceType(v string) *GatewayPatchBuilder {
	b.patch["service_type"] = v
	return b
}

func (b *GatewayPatchBuilder) Status(v string) *GatewayPatchBuilder {
	b.patch["status"] = v
	return b
}

func (b *GatewayPatchBuilder) SupervisorImage(v string) *GatewayPatchBuilder {
	b.patch["supervisor_image"] = v
	return b
}

func (b *GatewayPatchBuilder) TLSMode(v string) *GatewayPatchBuilder {
	b.patch["tls_mode"] = v
	return b
}


func (b *GatewayPatchBuilder) Build() map[string]any {
	return b.patch
}